pass current loopback user id into credentials functions.
This would greatly aid me in a current use case I have where Im storing the
actual credentials in redis and I look them up by user id. Or rather, I would
like to be able to look them up by user id but in order to do that I need to have
access to the current user id in the credentials functions.